feat: Remove cached nodes when finished (#15310)

+++ crates/polars-plan/src/logical_plan/alp/mod.rs
@@ -68,7 +68,7 @@ pub enum ALogicalPlan {
         // Unique ID.
         id: usize,
         /// How many hits the cache must be saved in memory.
-        cache_hits: usize,
+        cache_hits: u32,
     },
     Aggregate {
         input: Node,
